Move cmat_inds_names to qha_general_stress
